Explain xkcd does not longer send the image wihtin the email
All emails contains the xkcd images twice one for xkcd itself and one for the explain web page.
By now explain pictures are not fetched anymore.

Bugfix: The image url was written to the file without the prefix 'hash'. So every time the script
was run it sends the explain comics